new update

+++ Scripts/teiler-alt.sh
@@ -11,32 +11,45 @@ if [ -z "$XDG_RUNTIME_DIR" ]; then
     trap 'rm -rf $TMPDIR; exit' INT QUIT HUP TERM 0
 fi
 
-source /etc/teiler/config
+if [[ -d /etc/teiler ]]
+then
+  source /etc/teiler/teiler.conf
+fi
 
 if [[ ! -d $HOME/.config/teiler ]]; then
     mkdir $HOME/.config/teiler
 fi
 
 if [[ ! -f $HOME/.config/teiler/config ]]; then
-    cp /etc/teiler/config $HOME/.config/teiler/config
+    cp /etc/teiler/teiler.conf $HOME/.config/teiler/config
 fi
 
 export ext=${ext}; source $HOME/.config/teiler/config
+
+if [[ ! -d $HOME/.config/teiler/uploader ]]; then
+    mkdir $HOME/.config/teiler/uploader
+    if [[ -d /etc/teiler ]]
+    then
+      cp /etc/teiler/uploader/* $HOME/.config/teiler/uploader
+    fi
+fi
+
 
 if [[ ! -d $HOME/.config/teiler/profiles ]]; then
     mkdir $HOME/.config/teiler/profiles
 fi
 
-dir=$(pwd)
-cd /etc/teiler/profiles
-for i in *; do
+if [[ -d /etc/teiler ]]
+then
+  cd /etc/teiler/profiles
+  for i in *; do
     if [[ ! -f $HOME/.config/teiler/profiles/"$i" ]]; then
-        cp /etc/teiler/profiles/"$i" $HOME/.config/teiler/profiles
+      cp /etc/teiler/profiles/"$i" $HOME/.config/teiler/profiles
     else
-        cmp "$i" $HOME/.config/teiler/profiles/"$i" || cp -f /etc/teiler/profiles/"$i" $HOME/.config/teiler/profiles
+      cmp "$i" $HOME/.config/teiler/profiles/"$i" || cp -f /etc/teiler/profiles/"$i" $HOME/.config/teiler/profiles
     fi
-done
-cd ${dir}
+  done
+fi
 
 check_img_path () {
   if [[ ! -d $img_path ]]; then
@@ -68,56 +81,101 @@ check_vid_path () {
   fi
 }
 
+check_paste_path () {
+  if [[ ! -d $paste_path ]]; then
+    read -p "Paste path \"${paste_path}\" does not exist. Create it? (Y/n) " -n 1 -r
+    echo
+    if [[ $REPLY =~ ^[Yy]$ || $REPLY == "" ]]; then
+      mkdir -p "${paste_path}"
+    elif [[ $REPLY =~ ^[Nn]$ ]]; then
+      exit
+    else
+      echo "Please enter y or n"
+      check_paste_path
+    fi
+  fi
+}
+
 source "$HOME/.config/teiler/profiles/${profile}"
 
 vid_filemask="${vid_filemask}.${ext}"
-noupload=$save
+save=$noupload
 
 mainMenu () {
-    if [[ -f $XDG_RUNTIME_DIR/__teiler_cast_name ]]; then
-        filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
-    fi
-    isRecording && STATE_RECORDING="1 Stop Recording Screencast"
-    if [[ "$STATE_RECORDING" == "1 Stop Recording Screencast" ]]; then
-        menu=$(echo -e "< Exit\n---\n"$STATE_RECORDING"" | _rofi -dmenu -u 2 -p "teiler > ")
-        val=$?
-        if [[ $val -eq 1 ]]; then exit;
-        elif [[ $val -eq 0 ]]; then askPrompt;
-        fi
+  if [[ -f $XDG_RUNTIME_DIR/__teiler_cast_name ]]
+  then
+    filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
+  fi
 
-    elif [[ -z "$STATE_RECORDING" ]]; then
-        menu=$(echo -e "< Exit\n---\n1 [ Screenshots ]>\n2 [ Screencasts ]>\n3 [ History ]>\n---" | _rofi -dmenu -p "teiler > ")
-        val=$?
-        if [[ $menu == "1 [ Screenshots ]>" ]]; then screenshotMenu;
-        elif [[ $menu == "2 [ Screencasts ]>" ]]; then screencastMenu;
-        elif [[ $menu == "3 [ History ]>" ]]; then uploadMenu;
-        elif [[ $menu == "< Exit" ]]; then exit;
-        elif [[ $val -eq 1 ]]; then exit; fi
+  isRecording && STATE_RECORDING="2 Stop Recording Screencast"
+
+  if [[ "$STATE_RECORDING" == "2 Stop Recording Screencast" ]]
+  then
+    HELP="<span color='$help_color'>${upload}: Upload | ${save}: No Upload</span>"
+    menu=$(echo -e "< Exit\n---\n"$STATE_RECORDING"" | _rofi -dmenu -u 2 -p "teiler > ")
+    val=$?
+    if [[ $val -eq 1 ]]
+    then
+      exit
+    elif [[ $val -eq 0 ]]
+    then
+      askPrompt
     fi
+  elif [[ -z "$STATE_RECORDING" ]]
+  then
+    menu=$(echo -e "< Exit\n---\n1 [ Screenshots ]>\n2 [ Screencasts ]>\n3 [ History ]>\n---" | _rofi -dmenu -p "teiler > ")
+    val=$?
+    if [[ $val -eq 0 ]]
+    then
+      case "${menu}" in
+        "1 [ Screenshots ]>") screenshotMenu;;
+        "2 [ Screencasts ]>") askPrompt;;
+        "3 [ History ]>") openHistory;;
+        "< Exit") exit;;
+      esac
+    elif [[ $val -eq 1 ]]
+    then
+      exit
+    fi
+  fi
 }
 
 screenshotMenu () {
+    help_text=$(echo -e "Upload: ${upload} - Save: ${save} - Default: Enter" | column -s '-' -t)
+    line1=$(echo "${help_text}" | head -1)
+    HELP="<span color='$help_color'>${line1}</span>"
     export filename="${img_filemask}"
-    menu=$(echo -e "< Return to Main Menu\n---\n1 Monitor\n2 Fullscreen\n3 Area" | _rofi -dmenu -p "Screenshot > ")
+    menu=$(echo -e "< Return to Main Menu\n---\n1 Monitor\n2 Fullscreen\n3 Area" | _rofi -dmenu -mesg "${HELP}" -p "Screenshot > ")
     rofi_exit=$?
-    if [[ $menu == "1 Monitor" ]]; then
-        mode="fullscreen"
-        desc="Monitor"
-    elif [[ $menu == "2 Fullscreen" ]]; then
-        mode="fullscreenAll"
-        desc="Fullscreen"
-    elif [[ $menu == "3 Area" ]]; then
-        mode="area"
-        desc="Area"
-    fi
-    if [[ $menu == "< Return to Main Menu" ]]; then
-        mainMenu
-    fi
-    if [[ $rofi_exit -eq 1 ]]; then
-        exit
-    elif [[ $val -eq 10 || $val -eq 0 ]]; then
+    case "${menu}" in
+        "1 Monitor") mode="fullscreen"; desc="Monitor";;
+        "2 Fullscreen") mode="fullscreenAll"; desc="Monitor";;
+        "3 Area") mode="area"; desc="Area";;
+        "< Return to Main Menu") mainMenu;;
+      esac
+    if [[ $rofi_exit -eq 10 ]]; then
         delayPrompt
-        maimCmd delay $mode "${filename}" && notify-send -a "teiler" -t "1500" "teiler" "Screenshot taken" && exit
+        if [[ $delay == "0" ]]; then maimCmd nodelay $mode "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}" && exit
+        else maimCmd delay $mode "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"; notify-send -a "teiler" "teiler" "Screenshot saved" && exit
+        fi
+    elif [[ $rofi_exit -eq 11 ]]; then
+        delayPrompt
+        if [[ $delay == "0" ]]; then maimCmd nodelay $mode "${filename}" && exit
+        else maimCmd delay $mode "${filename}"; notify-send -a "teiler" "teiler" "Screenshot saved"
+        fi
+    elif [[ $rofi_exit -eq 1 ]]; then
+        exit
+    elif [[ $rofi_exit -eq 0 ]]; then
+        delayPrompt
+        if [[ $always_ul == "1" ]]; then
+            if [[ $delay == "0" ]]; then maimCmd nodelay $mode "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}" && exit
+            else maimCmd delay $mode "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"; notify-send -a "teiler" "teiler" "Screenshot saved" && exit
+            fi
+        elif [[ $always_ul == "0" ]]; then
+            if [[ $delay == "0" ]]; then maimCmd nodelay $mode "${filename}" && exit
+            else maimCmd delay $mode "${filename}"; notify-send -a "teiler" "teiler" "Screenshot saved"
+            fi
+        fi
     fi
 }
 
@@ -128,18 +186,23 @@ test_xrandr () {
 }
 
 screencastMenu () {
-    #filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
-    isRecording && STATE_RECORDING="1 Stop Recording Screencast"
-    if [[ "$STATE_RECORDING" == "1 Stop Recording Screencast" ]]; then
-        filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
-        menu=$(echo -e "< Exit\n---\n"$STATE_RECORDING"" | _rofi -dmenu  -p "teiler > ")
+    filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
+    isRecording && STATE_RECORDING="2 Stop Recording Screencast"
+    if [[ "$STATE_RECORDING" == "2 Stop Recording Screencast" ]]; then
+        HELP="<span color='$help_color'>${upload}: Upload | ${save}: No Upload</span>"
+        menu=$(echo -e "< Exit\n---\n"$STATE_RECORDING"" | _rofi -dmenu -u 2 -kb-custom-1 "${upload}" -kb-custom-2 "${save}" -mesg "${HELP}" -p "teiler > ")
+        val=$?
 
         if [[ "$menu" == "< Exit" ]]; then
             exit
         fi
-        if [[ $val -eq 10 ]]; then isRecording && stopRecording && sleep 2 && cd "${vid_path}"\
-         && rm -f $XDG_RUNTIME_DIR/__teiler_cast_name;
+        if [[ $val -eq 10 ]]; then isRecording && stopRecording && sleep 2 && cd "${vid_path}" && teiler_helper --upload video "${filename}" && rm -f $XDG_RUNTIME_DIR/__teiler_cast_name;
+        elif [[ $val -eq 11 ]]; then isRecording && stopRecording; rm -f $XDG_RUNTIME_DIR/__teiler_cast_name; notify-send -a "teiler" "teiler" "Screencast saved"; exit
         elif [[ $val -eq 1 ]]; then exit
+        elif [[ $val -eq 0 ]]; then
+            if [[ $always_ul ]]; then isRecording && stopRecording && sleep 2 && cd "${vid_path}" && teiler_helper --upload video "${filename}";
+            elif [[ $always_ul == "0" ]]; then isRecording && stopRecording; rm -f $XDG_RUNTIME_DIR/__teiler_cast_name; notify-send -a "teiler" "teiler" "Screencast saved"; exit
+            fi
         fi
 
     elif [[ -z "$STATE_RECORDING" ]]; then
@@ -155,42 +218,12 @@ screencastMenu () {
     fi
 }
 
-uploadMenu () {
+openHistory () {
     menu=$(echo -e "< Return to Main Menu\n---\n1 Images\n2 Videos" | _rofi -dmenu -p "History > ")
-    if [[ $menu == "1 Images" ]]; then imageMenu;
-    elif [[ $menu == "2 Videos" ]]; then videoMenu;
+    if [[ $menu == "1 Images" ]]; then qimgv "${img_path}";
+    elif [[ $menu == "2 Videos" ]]; then qimgv "${vid_path}";
     elif [[ $menu == "< Return to Main Menu" ]]; then mainMenu;
     elif [[ $menu == "" ]]; then exit
-    fi
-}
-
-imageMenu () {
-    cd "${img_path}"
-    HELP="<span color='$help_color'>${edit}: Edit</span>"
-    imagemenu=$(echo -e "< Return to Upload Menu\n---\n$(ls -1 -r)" | _rofi -dmenu -select "${entry}" -kb-custom-1 "${view}" -kb-custom-2 "${historyupload}" -kb-custom-3 "${edit}" -kb-custom-4 "${clip}" -mesg "${HELP}" -p "Choose > ")
-    val=$?
-
-    imagemenu2="${imagemenu%.*}"
-    if [[ "${imagemenu}" == "< Return to Upload Menu" ]]; then uploadMenu; fi
-
-    if [[ $val -eq 10 || $val -eq 0 ]]; then $viewer "${imagemenu}"; export entry="${imagemenu}"; imageMenu
-    elif [[ $val -eq 12 ]]; then
-        cp "${imagemenu}" "${imagemenu2}-mod.png"
-        $editor "${imagemenu2}-mod.png"
-        export entry="${imagemenu}"
-        imageMenu
-    elif [[ $val -eq 1 ]]; then exit;
-    fi
-}
-
-videoMenu () {
-        cd "${vid_path}"
-    videomenu=$(echo -e "< Return to Upload Menu\n---\n$(ls -1)" | _rofi -dmenu -select "${entry}" -p "Choose > ")
-    val=$?
-    videomenu2="${videomenu%.*}"
-    if [[ "${videomenu}" == "< Return to Upload Menu" ]]; then uploadMenu; fi
-    if [[ $val -eq 10 || $val -eq 0 ]]; then $player "${videomenu}"; export entry="${videomenu}"; videoMenu
-    elif [[ $val -eq 1 ]]; then exit
     fi
 }
 
@@ -202,7 +235,7 @@ stopRecording () {
         res=$(cat $XDG_RUNTIME_DIR/teiler_res)
     fi
     [[ -f $SCREENCAST_PIDFILE ]] && { pid=$(cat "$SCREENCAST_PIDFILE"); isRecording && kill "$pid"; rm "$SCREENCAST_PIDFILE"; }
-    notify-send -a "teiler" -t "1500" "teiler" "Stopped recording"
+    notify-send -a "teiler" "teiler" "Stopped recording"
     if [[ -z $rate ]]; then
         output=$(xininfo -name); xrandr --output "$output" --mode "$res"; rm -f $XDG_RUNTIME_DIR/teiler_res; return 0
     else
@@ -212,28 +245,21 @@ stopRecording () {
     fi
 }
 
-webmConvert () {
-    cd && cd "${vid_path}"
-   
-    # convert mkv (filename) to webm (ext2)
-    ffmpeg -i "${filename}" $options "${webm_name}"$ext2
-
-    #delete the mkv file
-    rm -f "${filename}"
-}
-
 slopCmd () {
     slopString=$(slop -b "$slop_border" -c "$slop_color")
     if [[ "$slopString" == *"Cancel=true"* ]]; then
         exit
- #   else eval "${slopString}"
     fi
     }
 
 ffmpegCmd () {
+    round() {
+        echo $(( ${1:?need one argument} / 2 * 2))
+    }
+
     xres=$(echo "$res" | awk -F 'x' '{ print $1 }')
     yres=$(echo "$res" | awk -F 'x' '{ print $2 }')
-  
+
     if [[ -n $res ]]; then
       res_now="$(xininfo -mon-size | sed 's/ /x/')"
     else
@@ -257,7 +283,7 @@ ffmpegCmd () {
         else ffmpeg -f x11grab ${border} -s $res -i $ffmpeg_display+$ffmpeg_offset $ffaudio $encopts "${vid_path}/${filename}" &
         fi
         echo "$!" > "$SCREENCAST_PIDFILE"
-        notify-send -a "teiler" -t "1500" "teiler" "Screencast started"
+        notify-send -a "teiler" "teiler" "Screencast started"
     elif [[ $1 == "area" ]]; then
         read -r X Y W H G ID < <(slop -f "%x %y %w %h %g %i")
         X=$(round $X)
@@ -265,29 +291,27 @@ ffmpegCmd () {
         W=$(round $W)
         H=$(round $H)
         isRecording && { notify "$time" 'Screencast already in progress'; echo "Already recording Screen"; exit 1; }
-        ffmpeg_display=$(echo $DISPLAY); ffmpeg_offset=$(echo $(xininfo -mon-x),$(xininfo -mon-y)); slopCmd
+        ffmpeg_display=$(echo $DISPLAY); ffmpeg_offset=$(echo $(xininfo -mon-x),$(xininfo -mon-y))
         [[ -f "${vid_path}/${filename}" ]] && rm "${vid_path}/${filename}"
         if [[ -z $ffaudio ]]; then ffmpeg -f x11grab ${border} -s "$W"x"$H" -i $ffmpeg_display+$X,$Y $rect_encopts "${vid_path}/${filename}" &
         else ffmpeg -f x11grab ${border} -s "$W"x"$H" -i $ffmpeg_display+$X,$Y $ffaudio $rect_encopts "${vid_path}/${filename}" &
         fi
         echo "$!" > "$SCREENCAST_PIDFILE"
-        notify-send -a "teiler" -t "1500" "teiler" "Screencast started"
+        notify-send -a "teiler" "teiler" "Screencast started"
     fi
 }
 
 maimCmd () {
     if [[ $hidecursor == "yes" ]]; then
-        cursor=""
+        cursor="--hidecursor"
     elif [[ $hidecursor == "no" ]]; then
-        cursor="--showcursor"
+        cursor=""
     fi
-
     if [[ $slop_color == "" ]]; then
       rect_color=""
     else
       rect_color="-c ${slop_color}"
     fi
-
     if [[ $slop_border == "" ]]; then
       rect_border=""
     else
@@ -295,9 +319,14 @@ maimCmd () {
     fi
 
     if [[ $1 == "delay" ]]; then
-        if [[ $2 == "area" ]]; then maim $cursor ${rect_border} ${rect_color} -s -d ${delay} "${img_path}/${3}";
+      if [[ $2 == "area" ]]; then maim $cursor ${rect_border} ${rect_color} -s -d ${delay} "${img_path}/${3}";
         elif [[ $2 == "fullscreen" ]]; then maim $cursor -g $(xininfo -mon-width)x$(xininfo -mon-height)+$(xininfo -mon-x)+$(xininfo -mon-y) -d ${delay} "${img_path}/${3}";
         elif [[ $2 == "fullscreenAll" ]]; then maim $cursor -d ${delay} "${img_path}/${3}"; fi
+    elif [[ $1 == "nodelay" ]]; then
+        if [[ $2 == "area" ]]; then maim $cursor ${rect_border} ${rect_color} -s "${img_path}/${3}";
+        elif [[ $2 == "active" ]]; then maim -i $(xdotool getactivewindow) "${img_path}/${3}";
+        elif [[ $2 == "fullscreen" ]]; then maim $cursor -g $(xininfo -mon-width)x$(xininfo -mon-height)+$(xininfo -mon-x)+$(xininfo -mon-y) "${img_path}/${3}";
+        elif [[ $2 == "fullscreenAll" ]]; then maim $cursor "${img_path}/${3}"; fi
     fi
 }
 
@@ -315,16 +344,23 @@ askPrompt () {
     else
         filename="$(cat $XDG_RUNTIME_DIR/__teiler_cast_name)"
         isRecording && stopRecording && sleep 2
-        rm -f $XDG_RUNTIME_DIR/__teiler_cast_name
-        webmConvert
+        ask=$(echo -e "1 Yes\n2 No" | _rofi -dmenu -p "Upload? > ")
+        if [[ $ask == "1 Yes" ]]; then cd "${vid_path}" && teiler_helper --upload video "${filename}" && rm -f $XDG_RUNTIME_DIR/__teiler_cast_name;
+        elif [[ $ask == "2 No" ]]; then rm -f $XDG_RUNTIME_DIR/__teiler_cast_name;
+        fi
     fi
 }
 
+clipCmd () {
+    (xclip -o) > "${paste_path}/${paste_filemask}"
+}
+
 delayPrompt () {
-    delay=$(echo -e "$(seq 1 10)" | _rofi -dmenu -p "Choose Delay > ")
-    if [[ $delay == "" ]]; then
-        exit
-    fi
+    delay=1
+    # delay=$(echo -e "$(seq 0 10)" | _rofi -dmenu -p "Choose Delay > ")
+    # if [[ $delay == "" ]]; then
+    #     exit
+    # fi
 }
 
 helpCmd () {
@@ -333,7 +369,8 @@ teiler - a rofi-driven screen{shot,cast} utility
 (C) Rasmus Steinke <rasi@xssn.at>
 --screenshot                               open screenshots menu
 --screencast                               open screencasts menu
---history {images,video}                   open history menus
+--history                                  open history menu
+--paste                                    upload text from clipboard
 --togglecast                               start/stop screencast
 --quick image {area,fullscreen,all,active} quickly create screenshot and upload
 --quick video {area,fullscreen}            quickly create screenshot and optionally upload
@@ -342,21 +379,20 @@ EOF
 }
 
 
+
 if [[ $1 == "--screenshot" ]]; then check_img_path; screenshotMenu
 elif [[ $1 == "--screencast" ]]; then check_vid_path; screencastMenu
+elif [[ $1 == "--paste" ]]; then check_paste_path; export filename="${paste_filemask}" && clipCmd && teiler_helper --upload text "${filename}"
 elif [[ $1 == "--togglecast" ]]; then check_vid_path; askPrompt
-elif [[ $1 == "--history" ]]; then
-    if [[ $2 == "images" ]]; then check_img_path; imageMenu
-    elif [[ $2 == "videos" ]]; then check_vid_path; videoMenu
-    fi
+elif [[ $1 == "--history" ]]; then openHistory;
 elif [[ $1 == "--quick" ]]; then
     if [[ $2 == "image" ]]; then
       check_img_path
       filename="${img_filemask}"
-      if [[ $3 == "area" ]]; then maimCmd nodelay area "${filename}" && cd "${img_path}"
-      elif [[ $3 == "active" ]]; then maimCmd nodelay active "${filename}" && cd "${img_path}"
-      elif [[ $3 == "fullscreen" ]]; then maimCmd nodelay fullscreen "${filename}" && cd "${img_path}"
-      elif [[ $3 == "all" ]]; then maimCmd nodelay fullscreenAll "${filename}" && cd "${img_path}"
+      if [[ $3 == "area" ]]; then maimCmd nodelay area "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"
+      elif [[ $3 == "active" ]]; then maimCmd nodelay active "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"
+      elif [[ $3 == "fullscreen" ]]; then maimCmd nodelay fullscreen "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"
+      elif [[ $3 == "all" ]]; then maimCmd nodelay fullscreenAll "${filename}" && cd "${img_path}" && teiler_helper --upload image "${filename}"
       fi
     elif [[ $2 == "video" ]]; then
       check_vid_path
@@ -379,4 +415,3 @@ elif [[ $1 == "--help" || $1 == "-h" ]]; then
 else
     check_img_path; mainMenu
 fi
-
